-subagent upgrade, separate table
