+++ ygg-chat/client/ygg-chat-r/package.json
@@ -1,7 +1,7 @@
 {
   "name": "graviton",
   "private": true,
-  "version": "0.2.2",
+  "version": "0.2.3",
   "description": "YggChat desktop client",
   "homepage": "https://yggchat.com",
   "author": {
