+++ ygg-chat/client/ygg-chat-r/src/features/chats/chatActions.ts
@@ -43,13 +43,6 @@ import { openStreamingWithPreFirstByteRetry } from './streamResilience'
 import { buildCompactionHistoryLines, buildCompactionWriteOpAppendix } from './compactionContext'
 import { persistToolResultsWithFallback } from './toolResultPersistence'
 // OpenAI OAuth is handled internally by OpenAIChatGPT module
-import {
-  getSubagentEnabledTools,
-  getSubagentMaxTurns,
-  isOrchestratorEnabled,
-  shouldUseGlobalAgentModelForSubagentDefault,
-} from '../../helpers/subagentToolSettings'
-import { loadAgentSettings } from '../../helpers/agentSettingsStorage'
 import { loadAutoCompactionEnabled } from '../../helpers/chatUiSettingsStorage'
 import { DEFAULT_COMPACTION_SYSTEM_PROMPT, loadProviderSettings } from '../../helpers/providerSettingsStorage'
 import { getDefaultBashTimeoutMs, getDefaultToolCallTimeoutMs } from '../../helpers/toolExecutionSettings'
@@ -74,6 +67,11 @@ import {
   type PlanClarificationAnswer,
   type PlanClarificationResult,
 } from './planToolTypes'
+import {
+  abortSubagentControllers,
+  executeSimpleSubagentCall,
+  executeSubagentCall,
+} from './subagentRuntime'
 
 // TODO: Import when conversations feature is available
 // import { conversationActions } from '../conversations'
@@ -1433,222 +1431,7 @@ const executeBrowseWebLocally = async (
   }
 }
 
-const DEFAULT_SUBAGENT_MODEL = 'openai/gpt-5.3-codex'
-
-const normalizeProviderSlug = (providerName: string | null | undefined): string =>
-  (providerName || '').toLowerCase().replace(/\s+/g, '')
-
-type SubagentInheritedProvider = 'openaichatgpt' | 'openrouter' | 'lmstudio' | 'zai' | 'bedrock'
-
-const resolveInheritedSubagentProvider = (
-  callerProviderName: string | null | undefined
-): SubagentInheritedProvider | undefined => {
-  const slug = normalizeProviderSlug(callerProviderName)
-  if (slug === 'openaichatgpt' || slug === 'openai(chatgpt)') return 'openaichatgpt'
-  if (slug === 'openrouter') return 'openrouter'
-  if (slug === 'lmstudio') return 'lmstudio'
-  if (slug === 'zai' || slug === 'z.ai' || slug === 'glm') return 'zai'
-  if (slug === 'bedrock' || slug === 'awsbedrock' || slug === 'aws-bedrock' || slug === 'amazonbedrock' || slug === 'amazon-bedrock') return 'bedrock'
-  return undefined
-}
-
-const resolveSubagentDefaults = async (
-  requestedModel: unknown,
-  callerProviderName?: string | null
-): Promise<{ model: string; provider?: SubagentInheritedProvider }> => {
-  const normalizedRequestedModel = typeof requestedModel === 'string' ? requestedModel.trim() : ''
-
-  let defaultModel = DEFAULT_SUBAGENT_MODEL
-  if (shouldUseGlobalAgentModelForSubagentDefault()) {
-    try {
-      const agentSettings = await loadAgentSettings()
-      if (typeof agentSettings.model === 'string' && agentSettings.model.trim().length > 0) {
-        defaultModel = agentSettings.model.trim()
-      }
-    } catch (error) {
-      console.warn('[subagent] Failed to load global agent model for defaulting:', error)
-    }
-  }
-
-  return {
-    model: normalizedRequestedModel || defaultModel,
-    provider: resolveInheritedSubagentProvider(callerProviderName),
-  }
-}
-
-const shouldUseCommunityLocalEphemeral = () => isCommunityMode && isElectronEnvironment
-
-const parseToolResultsFromContentBlocks = (contentBlocks: any): any[] => {
-  if (!Array.isArray(contentBlocks)) return []
-
-  return contentBlocks
-    .filter(block => block?.type === 'tool_result' && (block?.tool_use_id || block?.toolUseId))
-    .map(block => ({
-      tool_use_id: block.tool_use_id || block.toolUseId,
-      content:
-        typeof block.content === 'string' ? block.content : block.content == null ? '' : JSON.stringify(block.content),
-      is_error: Boolean(block.is_error ?? block.isError),
-      tool_name: typeof block.name === 'string' ? block.name : undefined,
-      input: block.input,
-    }))
-}
-
-/**
- * Simple subagent execution without tool calling (fallback when context not available)
- */
-const executeSimpleSubagentCall = async (
-  toolCall: any,
-  accessToken: string | null,
-  callerProviderName?: string | null
-): Promise<string> => {
-  const args = toolCall.arguments || {}
-  const { prompt, model, systemPrompt, maxTokens, temperature, response_format, responseFormat } = args
-  const effectiveResponseFormat = response_format ?? responseFormat
-
-  if (!prompt) {
-    throw new Error('Subagent requires a prompt')
-  }
-
-  const { model: resolvedModel, provider: resolvedProvider } = await resolveSubagentDefaults(model, callerProviderName)
-
-  try {
-    if (shouldUseCommunityLocalEphemeral()) {
-      const localPayload = await localApi.post<any>('/headless/ephemeral/chat', {
-        prompt,
-        provider: resolvedProvider,
-        model: resolvedModel,
-        maxTokens,
-        temperature: temperature ?? 0.7,
-        systemPrompt,
-        response_format: effectiveResponseFormat,
-      })
-
-      if (!localPayload?.success) {
-        throw new Error(localPayload?.error || 'Community local subagent generation failed')
-      }
-
-      const fullText = typeof localPayload?.message?.content === 'string' ? localPayload.message.content : ''
-      const reasoning = typeof localPayload?.reasoning === 'string' ? localPayload.reasoning : ''
-
-      if (reasoning) {
-        return `<thinking>\n${reasoning}\n</thinking>\n\n${fullText}`
-      }
-      return fullText || 'Subagent returned empty response'
-    }
-    const response = await createStreamingRequest('/generate/ephemeral', accessToken, {
-      method: 'POST',
-      headers: { 'Content-Type': 'application/json' },
-      body: JSON.stringify({
-        prompt,
-        provider: resolvedProvider,
-        model: resolvedModel,
-        maxTokens,
-        temperature: temperature ?? 0.7,
-        systemPrompt,
-        response_format: effectiveResponseFormat,
-      }),
-    })
-
-    if (!response.ok) {
-      const errorText = await response.text()
-      throw new Error(`Subagent generation failed: HTTP ${response.status}: ${errorText}`)
-    }
-
-    const reader = response.body?.getReader()
-    if (!reader) {
-      throw new Error('No response body from subagent')
-    }
-
-    const decoder = new TextDecoder()
-    let fullText = ''
-    let reasoning = ''
-    let sseBuffer = ''
-
-    while (true) {
-      const { done, value } = await reader.read()
-      if (done) break
-
-      const chunk = decoder.decode(value, { stream: true })
-      sseBuffer += chunk
-
-      const lines = sseBuffer.split('\n')
-      sseBuffer = lines.pop() || ''
-
-      for (const line of lines) {
-        if (line.startsWith('data: ')) {
-          const data = line.slice(6)
-          if (data === '[DONE]') continue
-          try {
-            const parsed = JSON.parse(data)
-            if (parsed.text) {
-              fullText += parsed.text
-            } else if (parsed.reasoning) {
-              reasoning += parsed.reasoning
-            }
-          } catch {
-            if (data.trim()) {
-              fullText += data
-            }
-          }
-        }
-      }
-    }
-
-    if (reasoning) {
-      return `<thinking>\n${reasoning}\n</thinking>\n\n${fullText}`
-    }
-    return fullText || 'Subagent returned empty response'
-  } catch (error) {
-    console.error('[executeSimpleSubagentCall] Error:', error)
-    throw error instanceof Error ? error : new Error(String(error))
-  }
-}
-
-/**
- * Convert ToolDefinition to server's expected format for openrouter
- * Server expects: { name, enabled, description, inputSchema }
- */
-const convertToServerToolFormat = (tool: ToolDefinition) => ({
-  name: tool.name,
-  enabled: true, // Tools sent to subagent are explicitly enabled
-  description: tool.description || '',
-  inputSchema: tool.inputSchema || { type: 'object', properties: {} },
-})
-
-const subagentAbortControllersByStream = new Map<string, Set<AbortController>>()
 const generationAbortControllersByStream = new Map<string, Set<AbortController>>()
-
-const registerSubagentAbortController = (streamId: string | null | undefined, controller: AbortController) => {
-  if (!streamId) return () => {}
-  let controllers = subagentAbortControllersByStream.get(streamId)
-  if (!controllers) {
-    controllers = new Set()
-    subagentAbortControllersByStream.set(streamId, controllers)
-  }
-  controllers.add(controller)
-  return () => {
-    const set = subagentAbortControllersByStream.get(streamId)
-    if (!set) return
-    set.delete(controller)
-    if (set.size === 0) subagentAbortControllersByStream.delete(streamId)
-  }
-}
-
-const abortSubagentControllers = (streamId?: string | null) => {
-  if (streamId) {
-    const controllers = subagentAbortControllersByStream.get(streamId)
-    if (controllers) {
-      controllers.forEach(controller => controller.abort())
-      subagentAbortControllersByStream.delete(streamId)
-    }
-    return
-  }
-
-  for (const controllers of subagentAbortControllersByStream.values()) {
-    controllers.forEach(controller => controller.abort())
-  }
-  subagentAbortControllersByStream.clear()
-}
 
 const registerGenerationAbortController = (streamId: string | null | undefined, controller: AbortController) => {
   if (!streamId) return () => {}
@@ -1680,817 +1463,6 @@ const abortGenerationControllers = (streamId?: string | null) => {
     controllers.forEach(controller => controller.abort())
   }
   generationAbortControllersByStream.clear()
-}
-
-/**
- * Get filtered tool definitions for subagent based on mode
- * Returns tools in server's expected format: { name, enabled, description, inputSchema }
- */
-const getSubagentToolDefinitions = (
-  orchestratorMode: boolean,
-  requestedTools: string[] | undefined
-): Array<{ name: string; enabled: boolean; description: string; inputSchema: any }> => {
-  // Check if orchestrator is globally enabled
-  if (!isOrchestratorEnabled()) {
-    // Orchestrator disabled - subagent cannot use any tools
-    return []
-  }
-
-  const allTools = getAllTools()
-
-  // Always exclude 'subagent' to prevent recursion
-  const excludedTools = new Set(['subagent'])
-
-  let allowedToolNames: Set<string>
-
-  if (orchestratorMode && requestedTools?.length) {
-    // Orchestrator mode: use requested tools (intersection with available)
-    allowedToolNames = new Set(requestedTools.filter(name => !excludedTools.has(name)))
-  } else {
-    // Pre-configured mode: use localStorage settings
-    const configuredTools = getSubagentEnabledTools()
-    allowedToolNames = new Set(configuredTools.filter(name => !excludedTools.has(name)))
-  }
-
-  // Filter and convert to OpenAI format
-  // When orchestratorMode=true with explicit requestedTools, bypass the t.enabled check
-  // This allows the model to request any tool regardless of global enabled state
-  const bypassEnabledCheck = !!(orchestratorMode && requestedTools?.length)
-
-  return allTools
-    .filter(t => {
-      const passesEnabledCheck = bypassEnabledCheck ? true : t.enabled
-      return passesEnabledCheck && allowedToolNames.has(t.name) && !excludedTools.has(t.name)
-    })
-    .map(convertToServerToolFormat)
-}
-
-const parseMaybeJson = <T>(value: any, fallback: T): T => {
-  if (value == null) return fallback
-  if (typeof value !== 'string') return value as T
-
-  try {
-    return JSON.parse(value) as T
-  } catch {
-    return fallback
-  }
-}
-
-const normalizeToolCallsForSubagentHistory = (rawToolCalls: any[]): any[] => {
-  return rawToolCalls
-    .map(tc => {
-      const id = tc?.id
-      const name =
-        typeof tc?.name === 'string' ? tc.name : typeof tc?.function?.name === 'string' ? tc.function.name : ''
-
-      const rawArgs = tc?.arguments ?? tc?.input ?? tc?.function?.arguments ?? {}
-      const argsString = typeof rawArgs === 'string' ? rawArgs : JSON.stringify(rawArgs)
-
-      if (!id || !name) return null
-
-      return {
-        id,
-        type: 'function',
-        function: { name, arguments: argsString },
-      }
-    })
-    .filter(Boolean)
-}
-
-const getCachedSubagentSessionMessages = (
-  queryClient: QueryClient | null | undefined,
-  conversationId: string,
-  sessionId: string
-): Message[] => {
-  const cached = queryClient?.getQueryData<{ messages: Message[]; tree: any }>([
-    'conversations',
-    conversationId,
-    'messages',
-  ])
-
-  const messages = Array.isArray(cached?.messages) ? cached.messages : []
-
-  return messages
-    .filter(
-      message =>
-        String(message.conversation_id) === String(conversationId) &&
-        message.role === 'ex_agent' &&
-        message.ex_agent_type === 'subagent' &&
-        message.ex_agent_session_id === sessionId
-    )
-    .sort((a, b) => new Date(a.created_at || 0).getTime() - new Date(b.created_at || 0).getTime())
-}
-
-const isSubagentUserPromptMessage = (_message: Message, blocks: any[], index: number): boolean => {
-  if (index === 0) return true
-  if (!Array.isArray(blocks)) return false
-  return blocks.some(block => block?.type === 'text' && block?.subagent_role === 'user_prompt')
-}
-
-const reconstructSubagentConversationHistory = (messages: Message[]): any[] => {
-  const history: any[] = []
-
-  messages.forEach((message, index) => {
-    const content = message.content || ''
-    const blocks = parseMaybeJson<any[]>(message.content_blocks, [])
-    const rawToolCalls = parseMaybeJson<any[]>(message.tool_calls, [])
-    const toolCalls = Array.isArray(rawToolCalls) ? rawToolCalls : []
-
-    if (isSubagentUserPromptMessage(message, blocks, index)) {
-      history.push({ role: 'user', content })
-      return
-    }
-
-    const toolCallsForHistory = toolCalls.length > 0 ? normalizeToolCallsForSubagentHistory(toolCalls) : []
-
-    history.push({
-      role: 'assistant',
-      content,
-      ...(toolCallsForHistory.length > 0 ? { tool_calls: toolCallsForHistory } : {}),
-    })
-
-    if (Array.isArray(blocks)) {
-      for (const block of blocks) {
-        if (block?.type !== 'tool_result') continue
-        history.push({
-          role: 'tool',
-          tool_call_id: block.tool_use_id,
-          content: typeof block.content === 'string' ? block.content : JSON.stringify(block.content ?? ''),
-        })
-      }
-    }
-  })
-
-  return history
-}
-
-/**
- * Execute subagent tool with full agentic capabilities.
- * Supports multi-turn tool execution, message persistence, and configurable tool access.
- */
-const executeSubagentCall = async (
-  toolCall: any,
-  accessToken: string | null,
-  context: {
-    dispatch: any
-    getState: () => RootState
-    conversationId: string
-    parentMessageId: string
-    streamId?: string
-    rootPath: string | null
-    callerProvider?: string | null
-    queryClient?: QueryClient | null
-  }
-): Promise<string> => {
-  const args = toolCall.arguments || {}
-  const {
-    prompt,
-    model,
-    systemPrompt,
-    maxTokens,
-    temperature,
-    response_format,
-    responseFormat,
-    orchestratorMode = false,
-    tools: requestedTools,
-    inheritAutoApprove = true,
-    sessionId,
-    resume = false,
-  } = args
-  const effectiveResponseFormat = response_format ?? responseFormat
-
-  if (!prompt) {
-    throw new Error('Subagent requires a prompt')
-  }
-
-  const { dispatch, getState, conversationId, parentMessageId, rootPath } = context
-  const streamId = context.streamId
-  const state = getState()
-  const callerProviderName = context.callerProvider ?? state.chat.providerState.currentProvider
-  const { model: resolvedModel, provider: resolvedProvider } = await resolveSubagentDefaults(model, callerProviderName)
-
-  // Get filtered tool definitions for this subagent
-  const subagentTools = getSubagentToolDefinitions(orchestratorMode, requestedTools)
-
-  const requestedSessionId = typeof sessionId === 'string' && sessionId.trim().length > 0 ? sessionId.trim() : null
-  const shouldResume = resume === true && !!requestedSessionId
-  const subagentSessionId = shouldResume ? requestedSessionId! : uuidv4()
-  let hadAnyToolActivity = false
-
-  const priorMessages = shouldResume
-    ? getCachedSubagentSessionMessages(context.queryClient, conversationId, subagentSessionId)
-    : []
-
-  if (shouldResume && priorMessages.length === 0) {
-    throw new Error(
-      `Cannot resume subagent session ${subagentSessionId}: no cached messages found. Open or reload this conversation first, then retry.`
-    )
-  }
-
-  const conversationHistory: any[] = shouldResume
-    ? [...reconstructSubagentConversationHistory(priorMessages), { role: 'user', content: prompt }]
-    : [{ role: 'user', content: prompt }]
-
-  // Persist subagent user prompt message to local storage
-  const promptMessageId = uuidv4()
-  const storageMode = state.conversations.items.find(c => c.id === conversationId)?.storage_mode
-  const isLocalMode = shouldUseLocalApi(storageMode)
-  const lastPriorMessage = priorMessages[priorMessages.length - 1] ?? null
-  const promptParentId = shouldResume && lastPriorMessage ? String(lastPriorMessage.id) : parentMessageId
-
-  if (isLocalMode) {
-    try {
-      await localApi.post('/sync/message', {
-        id: promptMessageId,
-        conversation_id: conversationId,
-        parent_id: promptParentId,
-        role: 'ex_agent',
-        content: prompt,
-        ex_agent_type: 'subagent',
-        ex_agent_session_id: subagentSessionId,
-        content_blocks: JSON.stringify([{ type: 'text', content: prompt, subagent_role: 'user_prompt' }]),
-        created_at: new Date().toISOString(),
-      })
-    } catch (err) {
-      console.warn('[subagent] Failed to persist prompt message:', err)
-    }
-  }
-
-  let finalResponse = ''
-  let lastAssistantMessageId = promptMessageId
-
-  const subagentAbortController = new AbortController()
-  const unregisterAbortController = registerSubagentAbortController(streamId, subagentAbortController)
-  const isStreamActive = () => {
-    if (!streamId) return true
-    return getState().chat.streaming.byId[streamId]?.active ?? false
-  }
-
-  // Agentic loop with user-configurable hard limit
-  const subagentMaxTurns = getSubagentMaxTurns()
-  let shouldContinue = true
-  try {
-    for (let turn = 0; turn < subagentMaxTurns && shouldContinue; turn++) {
-      if (!isStreamActive()) {
-        subagentAbortController.abort()
-        throw new Error('Subagent aborted')
-      }
-
-      // Call ephemeral endpoint with tools and conversation history
-      const requestBody = {
-        messages: conversationHistory,
-        provider: resolvedProvider,
-        model: resolvedModel,
-        maxTokens,
-        temperature: temperature ?? 0.7,
-        systemPrompt,
-        response_format: effectiveResponseFormat,
-        tools: subagentTools.length > 0 ? subagentTools : undefined,
-      }
-
-      let response: Response
-      if (shouldUseCommunityLocalEphemeral()) {
-        const localPayload = await localApi.post<any>('/headless/ephemeral/chat', requestBody)
-        if (!localPayload?.success) {
-          throw new Error(localPayload?.error || 'Community local subagent generation failed')
-        }
-
-        const syntheticEvents: string[] = []
-        if (typeof localPayload?.reasoning === 'string' && localPayload.reasoning.length > 0) {
-          syntheticEvents.push(
-            `data: ${JSON.stringify({ type: 'chunk', part: 'reasoning', delta: localPayload.reasoning })}`
-          )
-        }
-        if (typeof localPayload?.message?.content === 'string' && localPayload.message.content.length > 0) {
-          syntheticEvents.push(
-            `data: ${JSON.stringify({ type: 'chunk', part: 'text', delta: localPayload.message.content })}`
-          )
-        }
-
-        const localToolCalls = Array.isArray(localPayload?.toolCalls) ? localPayload.toolCalls : []
-        for (const toolCall of localToolCalls) {
-          syntheticEvents.push(`data: ${JSON.stringify({ type: 'tool_call', toolCall })}`)
-        }
-
-        const localToolResults = parseToolResultsFromContentBlocks(localPayload?.contentBlocks)
-        for (const toolResult of localToolResults) {
-          syntheticEvents.push(`data: ${JSON.stringify({ type: 'tool_result', toolResult })}`)
-        }
-
-        syntheticEvents.push('data: [DONE]')
-        response = new Response(`${syntheticEvents.join('\n')}\n`, {
-          status: 200,
-          headers: { 'Content-Type': 'text/event-stream' },
-        })
-      } else {
-        response = await createStreamingRequest('/generate/ephemeral', accessToken, {
-          method: 'POST',
-          headers: { 'Content-Type': 'application/json' },
-          body: JSON.stringify(requestBody),
-          signal: subagentAbortController.signal,
-        })
-      }
-
-      if (!response.ok) {
-        const errorText = await response.text()
-        throw new Error(`Subagent generation failed: HTTP ${response.status}: ${errorText}`)
-      }
-
-      const reader = response.body?.getReader()
-      if (!reader) {
-        throw new Error('No response body from subagent')
-      }
-
-      const decoder = new TextDecoder()
-      let turnText = ''
-      let turnReasoning = ''
-      const turnToolCalls: any[] = []
-      const serverToolResults: any[] = [] // Track tool results from server-executed tools (e.g., brave_search)
-      const seenToolCallIds = new Set<string>()
-      const seenToolResultIds = new Set<string>()
-      let sseBuffer = ''
-
-      const appendText = (value?: any) => {
-        if (typeof value === 'string' && value.length > 0) {
-          turnText += value
-        }
-      }
-
-      const appendReasoning = (value?: any) => {
-        if (typeof value === 'string' && value.length > 0) {
-          turnReasoning += value
-        }
-      }
-
-      const addToolCall = (toolCall?: any) => {
-        if (!toolCall) return
-        if (toolCall.id && seenToolCallIds.has(toolCall.id)) return
-        if (toolCall.id) seenToolCallIds.add(toolCall.id)
-        turnToolCalls.push(toolCall)
-      }
-
-      const addToolResult = (toolResult?: any) => {
-        if (!toolResult) return
-        if (toolResult.tool_use_id && seenToolResultIds.has(toolResult.tool_use_id)) return
-        if (toolResult.tool_use_id) seenToolResultIds.add(toolResult.tool_use_id)
-        serverToolResults.push(toolResult)
-      }
-
-      const buildContentBlocks = (toolResults: any[], includeToolUsesForResults = false) => {
-        const blocks: any[] = []
-        if (turnReasoning) {
-          blocks.push({ type: 'thinking', content: turnReasoning })
-        }
-        if (turnText) {
-          blocks.push({ type: 'text', content: turnText })
-        }
-        for (const tc of turnToolCalls) {
-          blocks.push({ type: 'tool_use', id: tc.id, name: tc.name, input: tc.arguments || tc.input })
-        }
-        if (includeToolUsesForResults && toolResults.length > 0) {
-          const existingToolUseIds = new Set(turnToolCalls.map(tc => tc.id))
-          for (const tr of toolResults) {
-            if (tr?.tool_use_id && !existingToolUseIds.has(tr.tool_use_id) && tr.tool_name) {
-              blocks.push({
-                type: 'tool_use',
-                id: tr.tool_use_id,
-                name: tr.tool_name,
-                input: tr.input ?? tr.args ?? tr.arguments,
-              })
-            }
-          }
-        }
-        for (const tr of toolResults) {
-          blocks.push({
-            type: 'tool_result',
-            tool_use_id: tr.tool_use_id,
-            content: tr.content,
-            is_error: tr.is_error,
-          })
-        }
-        return blocks
-      }
-
-      // Parse SSE stream
-      while (true) {
-        if (!isStreamActive()) {
-          subagentAbortController.abort()
-          throw new Error('Subagent aborted')
-        }
-
-        const { done, value } = await reader.read()
-        if (done) break
-
-        const chunk = decoder.decode(value, { stream: true })
-        sseBuffer += chunk
-
-        const lines = sseBuffer.split('\n')
-        sseBuffer = lines.pop() || ''
-
-        for (const line of lines) {
-          if (line.startsWith('data: ')) {
-            const data = line.slice(6)
-            if (data === '[DONE]') continue
-            try {
-              const parsed = JSON.parse(data)
-              if (parsed?.type === 'chunk' && parsed?.part) {
-                if (parsed.part === 'text') {
-                  appendText(parsed.delta ?? parsed.content ?? parsed.text)
-                } else if (parsed.part === 'reasoning') {
-                  appendReasoning(parsed.delta ?? parsed.reasoning)
-                } else if (parsed.part === 'tool_call') {
-                  addToolCall(parsed.toolCall)
-                } else if (parsed.part === 'tool_result') {
-                  addToolResult(parsed.toolResult)
-                }
-              } else if (parsed?.type === 'tool_call') {
-                addToolCall(parsed.toolCall || parsed.tool_call || parsed)
-              } else if (parsed?.type === 'tool_result') {
-                addToolResult(parsed.toolResult || parsed.tool_result || parsed)
-              } else if (parsed?.toolCall) {
-                addToolCall(parsed.toolCall)
-              } else if (parsed?.toolResult) {
-                // Server executed a server-only tool (e.g., brave_search)
-                addToolResult(parsed.toolResult)
-              } else if (parsed?.reasoning) {
-                appendReasoning(parsed.reasoning)
-              } else if (parsed?.text) {
-                appendText(parsed.text)
-              } else if (parsed?.delta) {
-                appendText(parsed.delta)
-              } else if (parsed?.content) {
-                appendText(parsed.content)
-              }
-            } catch {
-              if (data.trim()) {
-                turnText += data
-              }
-            }
-          }
-        }
-      }
-
-      // Create assistant message for this turn
-      const assistantMessageId = uuidv4()
-      const assistantMessage: any = {
-        id: assistantMessageId,
-        conversation_id: conversationId,
-        parent_id: lastAssistantMessageId,
-        role: 'ex_agent',
-        content: turnText,
-        ex_agent_type: 'subagent',
-        ex_agent_session_id: subagentSessionId,
-        tool_calls: turnToolCalls.length > 0 ? JSON.stringify(turnToolCalls) : null,
-        thinking_block: turnReasoning || null,
-        created_at: new Date().toISOString(),
-      }
-
-      // If no client tool calls AND no server tool results, this is the final response
-      if (turnToolCalls.length === 0 && serverToolResults.length === 0) {
-        finalResponse = turnReasoning ? `<thinking>\n${turnReasoning}\n</thinking>\n\n${turnText}` : turnText
-
-        // Build content_blocks for final message (text only, no tool calls/results)
-        const finalContentBlocks = buildContentBlocks([])
-        assistantMessage.content_blocks = JSON.stringify(finalContentBlocks)
-
-        // Persist final assistant message
-        if (isLocalMode) {
-          try {
-            await localApi.post('/sync/message', assistantMessage)
-          } catch (err) {
-            console.warn('[subagent] Failed to persist final assistant message:', err)
-          }
-        }
-        break
-      }
-
-      // If server executed tools (e.g., brave_search) but no client tools needed,
-      // we need to add the server results to conversation history and continue the loop
-      // so the model can process the results
-      if (turnToolCalls.length === 0 && serverToolResults.length > 0) {
-        const contentBlocks = buildContentBlocks(serverToolResults, true)
-        assistantMessage.content_blocks = JSON.stringify(contentBlocks)
-
-        if (isLocalMode) {
-          try {
-            await localApi.post('/sync/message', assistantMessage)
-          } catch (err) {
-            console.warn('[subagent] Failed to persist assistant message (server tools only):', err)
-          }
-        }
-
-        if (turnText || turnReasoning) {
-          conversationHistory.push({
-            role: 'assistant',
-            content: turnText || '',
-          })
-        }
-
-        hadAnyToolActivity = hadAnyToolActivity || serverToolResults.length > 0
-
-        // Add server tool results to conversation history
-        // The server already added the assistant message with tool_calls and executed them
-        // We just need to add the tool results so the model can continue
-        for (const tr of serverToolResults) {
-          conversationHistory.push({
-            role: 'tool',
-            tool_call_id: tr.tool_use_id,
-            content: tr.content,
-          })
-        }
-
-        lastAssistantMessageId = assistantMessageId
-        finalResponse = turnText
-
-        // Continue to next turn - model will process the tool results
-        continue
-      }
-
-      // Process client tool calls
-      const toolResults: any[] = []
-
-      // Build a set of tool IDs that were already executed by the server
-      const serverExecutedToolIds = new Set(serverToolResults.map(tr => tr.tool_use_id))
-      hadAnyToolActivity = hadAnyToolActivity || serverToolResults.length > 0
-
-      // Filter out tool calls that were already executed by the server
-      const clientToolCalls = turnToolCalls.filter(tc => !serverExecutedToolIds.has(tc.id))
-
-      for (let i = 0; i < clientToolCalls.length; i++) {
-        const tc = clientToolCalls[i]
-
-        // Skip nested subagent calls
-        if (tc.name === 'subagent') {
-          toolResults.push({
-            tool_use_id: tc.id,
-            content: 'Error: Nested subagent calls are not allowed.',
-            is_error: true,
-          })
-          hadAnyToolActivity = true
-          continue
-        }
-
-        try {
-          // Read live settings per tool call so toggles apply mid-stream.
-          const liveState = getState()
-          const parentAutoApprove = liveState.chat.toolAutoApprove
-          const liveOperationMode = liveState.chat.operationMode
-          const shouldAutoApprove = inheritAutoApprove && parentAutoApprove
-
-          let result: string
-          if (shouldAutoApprove) {
-            // Execute directly without permission check
-            result = await executeLocalTool(tc, rootPath, liveOperationMode, {
-              conversationId,
-              messageId: assistantMessageId,
-              accessToken,
-              queryClient: context.queryClient ?? null,
-            })
-          } else {
-            // Show permission dialog
-            result = await executeToolWithPermissionCheck(dispatch, getState, tc, rootPath, liveOperationMode, {
-              conversationId,
-              messageId: assistantMessageId,
-              accessToken,
-              queryClient: context.queryClient ?? null,
-            })
-          }
-
-          toolResults.push({
-            tool_use_id: tc.id,
-            content: result,
-            is_error: false,
-          })
-          hadAnyToolActivity = true
-        } catch (err) {
-          const errorMsg = err instanceof Error ? err.message : String(err)
-          toolResults.push({
-            tool_use_id: tc.id,
-            content: `Error: ${errorMsg}`,
-            is_error: true,
-          })
-          hadAnyToolActivity = true
-        }
-      }
-
-      // Combine all tool results (server-executed + client-executed)
-      const allToolResults = [...serverToolResults, ...toolResults]
-
-      // Build content_blocks with tool results
-      const contentBlocks = buildContentBlocks(allToolResults)
-      assistantMessage.content_blocks = JSON.stringify(contentBlocks)
-
-      // Persist assistant message with tool calls and results
-      if (isLocalMode) {
-        try {
-          await localApi.post('/sync/message', assistantMessage)
-        } catch (err) {
-          console.warn('[subagent] Failed to persist assistant message:', err)
-        }
-      }
-
-      // Add assistant message with tool_calls to conversation history
-      const toolCallsForHistory = turnToolCalls
-        .map(tc => ({
-          id: tc?.id,
-          type: 'function',
-          function: {
-            name:
-              typeof tc?.name === 'string' ? tc.name : typeof tc?.function?.name === 'string' ? tc.function.name : '',
-            arguments: JSON.stringify(tc.arguments || tc.input || tc?.function?.arguments || {}),
-          },
-        }))
-        .filter(
-          (tc): tc is { id: string; type: 'function'; function: { name: string; arguments: string } } =>
-            typeof tc.id === 'string' &&
-            tc.id.length > 0 &&
-            typeof tc.function.name === 'string' &&
-            tc.function.name.length > 0
-        )
-
-      conversationHistory.push({
-        role: 'assistant',
-        content: turnText || '',
-        tool_calls: toolCallsForHistory,
-      })
-
-      // Add each tool result as a separate 'tool' message (same format as createToolResultMessage)
-      // Include both server-executed and client-executed tool results
-      for (const tr of allToolResults) {
-        conversationHistory.push({
-          role: 'tool',
-          tool_call_id: tr.tool_use_id,
-          content: tr.content,
-        })
-      }
-
-      lastAssistantMessageId = assistantMessageId
-      finalResponse = turnText
-    }
-  } catch (error) {
-    if (subagentAbortController.signal.aborted) {
-      throw new Error('Subagent aborted')
-    }
-    console.error('[subagent] Error in subagent execution:', error)
-    throw error
-  } finally {
-    unregisterAbortController()
-  }
-
-  // Deterministic finalization: if tools ran but no text response, force one more turn
-  const hasFinalText = typeof finalResponse === 'string' && finalResponse.trim().length > 0
-
-  if (hadAnyToolActivity && !hasFinalText && !subagentAbortController.signal.aborted) {
-    if (!isStreamActive()) {
-      subagentAbortController.abort()
-      throw new Error('Subagent aborted')
-    }
-
-    // Append a finalization instruction in subagent-only context
-    conversationHistory.push({
-      role: 'user',
-      content:
-        'Summarize the tool results above and provide the final answer. Do not call tools. Be concise and complete.',
-    })
-
-    const finalizeRequestBody = {
-      messages: conversationHistory,
-      provider: resolvedProvider,
-      model: resolvedModel,
-      maxTokens,
-      temperature: temperature ?? 0.3,
-      systemPrompt,
-      tools: undefined, // Force no tool calls for finalization
-    }
-
-    let finalizeResponse: Response
-    if (shouldUseCommunityLocalEphemeral()) {
-      const localPayload = await localApi.post<any>('/headless/ephemeral/chat', finalizeRequestBody)
-      if (!localPayload?.success) {
-        throw new Error(localPayload?.error || 'Community local subagent finalization failed')
-      }
-
-      const syntheticEvents: string[] = []
-      if (typeof localPayload?.reasoning === 'string' && localPayload.reasoning.length > 0) {
-        syntheticEvents.push(
-          `data: ${JSON.stringify({ type: 'chunk', part: 'reasoning', delta: localPayload.reasoning })}`
-        )
-      }
-      if (typeof localPayload?.message?.content === 'string' && localPayload.message.content.length > 0) {
-        syntheticEvents.push(
-          `data: ${JSON.stringify({ type: 'chunk', part: 'text', delta: localPayload.message.content })}`
-        )
-      }
-      syntheticEvents.push('data: [DONE]')
-
-      finalizeResponse = new Response(`${syntheticEvents.join('\n')}\n`, {
-        status: 200,
-        headers: { 'Content-Type': 'text/event-stream' },
-      })
-    } else {
-      finalizeResponse = await createStreamingRequest('/generate/ephemeral', accessToken, {
-        method: 'POST',
-        headers: { 'Content-Type': 'application/json' },
-        body: JSON.stringify(finalizeRequestBody),
-        signal: subagentAbortController.signal,
-      })
-    }
-
-    if (!finalizeResponse.ok) {
-      const errorText = await finalizeResponse.text()
-      throw new Error(`Subagent finalization failed: HTTP ${finalizeResponse.status}: ${errorText}`)
-    }
-
-    const finalizeReader = finalizeResponse.body?.getReader()
-    if (!finalizeReader) {
-      throw new Error('No response body from subagent finalization')
-    }
-
-    const finalizeDecoder = new TextDecoder()
-    let finalizeText = ''
-    let finalizeReasoning = ''
-    let finalizeBuffer = ''
-
-    while (true) {
-      if (!isStreamActive()) {
-        subagentAbortController.abort()
-        throw new Error('Subagent aborted')
-      }
-
-      const { done, value } = await finalizeReader.read()
-      if (done) break
-
-      const chunk = finalizeDecoder.decode(value, { stream: true })
-      finalizeBuffer += chunk
-
-      const lines = finalizeBuffer.split('\n')
-      finalizeBuffer = lines.pop() || ''
-
-      for (const line of lines) {
-        if (!line.startsWith('data: ')) continue
-        const data = line.slice(6)
-        if (data === '[DONE]') continue
-        try {
-          const parsed = JSON.parse(data)
-          if (parsed?.type === 'chunk' && parsed?.part) {
-            if (parsed.part === 'text') {
-              const delta = parsed.delta ?? parsed.content ?? parsed.text
-              if (typeof delta === 'string') finalizeText += delta
-            } else if (parsed.part === 'reasoning') {
-              const delta = parsed.delta ?? parsed.reasoning
-              if (typeof delta === 'string') finalizeReasoning += delta
-            }
-          } else if (parsed?.text) {
-            if (typeof parsed.text === 'string') finalizeText += parsed.text
-          } else if (parsed?.reasoning) {
-            if (typeof parsed.reasoning === 'string') finalizeReasoning += parsed.reasoning
-          } else if (parsed?.delta) {
-            if (typeof parsed.delta === 'string') finalizeText += parsed.delta
-          } else if (parsed?.content) {
-            if (typeof parsed.content === 'string') finalizeText += parsed.content
-          }
-        } catch {
-          if (data.trim()) {
-            finalizeText += data
-          }
-        }
-      }
-    }
-
-    const finalizeMessageId = uuidv4()
-    const finalizeMessage: any = {
-      id: finalizeMessageId,
-      conversation_id: conversationId,
-      parent_id: lastAssistantMessageId,
-      role: 'ex_agent',
-      content: finalizeText,
-      ex_agent_type: 'subagent',
-      ex_agent_session_id: subagentSessionId,
-      thinking_block: finalizeReasoning || null,
-      created_at: new Date().toISOString(),
-    }
-
-    const finalizeBlocks: any[] = []
-    if (finalizeReasoning) finalizeBlocks.push({ type: 'thinking', content: finalizeReasoning })
-    if (finalizeText) finalizeBlocks.push({ type: 'text', content: finalizeText })
-    finalizeMessage.content_blocks = JSON.stringify(finalizeBlocks)
-
-    if (isLocalMode) {
-      try {
-        await localApi.post('/sync/message', finalizeMessage)
-      } catch (err) {
-        console.warn('[subagent] Failed to persist finalization message:', err)
-      }
-    }
-
-    finalResponse = finalizeReasoning
-      ? `<thinking>\n${finalizeReasoning}\n</thinking>\n\n${finalizeText}`
-      : finalizeText
-    lastAssistantMessageId = finalizeMessageId
-  }
-
-  return finalResponse || 'No response generated'
-  // legacy summary removed\n\n${finalResponse || 'No response generated'}\n\n---\nTurns: ${turnsUsed}/${maxTurns} | Tool calls: ${totalToolCallsUsed}/${maxToolCalls} | Tools: ${toolSummary}`
 }
 
 /**
@@ -2536,6 +1508,8 @@ export const executeLocalTool = async (
       rootPath,
       callerProvider: context.callerProvider,
       queryClient: context.queryClient ?? null,
+      executeLocalTool,
+      executeToolWithPermissionCheck,
     })
   }
